requirements.txt
requests
lxml
chromadb
openai
python-dotenv
gradio
fastapi
uvicorn
jinja2
pydantic